backend_v2/ml_service/pyproject.toml
[project]
name = "skincare-ml-service"
version = "0.1.0"
description = "PyTorch/ONNX skin analysis inference sidecar for backend_v2"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "numpy>=1.26.0",
    "pillow>=10.4.0",
    "opencv-python-headless>=4.10.0",
    "onnxruntime>=1.19.0",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]
# Heavy deps only needed when training / exporting models (run in Colab, not in prod container).
train = [
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "timm>=1.0.9",
    "scikit-learn>=1.5.0",
    "pandas>=2.2.0",
    "matplotlib>=3.9.0",
    "tqdm>=4.66.0",
    "grad-cam>=1.5.4",
    "onnx>=1.17.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["models*", "tests*", "scripts*", ".venv*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
